Language.Haskell.TH.Desugar.Lift

Description

Defines Lift instances for the desugared language. This is defined in a separate module because it also must define Lift instances for several TH types, which are orphans and may want another definition downstream.

Orphan instances

Lift DClause

Instance details

Methods

lift :: Quote m => DClause -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DClause -> Code m DClause

Lift DCon

Instance details

Methods

lift :: Quote m => DCon -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DCon -> Code m DCon

Lift DConFields

Instance details

Methods

lift :: Quote m => DConFields -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DConFields -> Code m DConFields

Lift DDec

Instance details

Methods

lift :: Quote m => DDec -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DDec -> Code m DDec

Lift DDerivClause

Instance details

Methods

lift :: Quote m => DDerivClause -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DDerivClause -> Code m DDerivClause

Lift DDerivStrategy

Instance details

Methods

lift :: Quote m => DDerivStrategy -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DDerivStrategy -> Code m DDerivStrategy

Lift DExp

Instance details

Methods

lift :: Quote m => DExp -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DExp -> Code m DExp

Lift DFamilyResultSig

Instance details

Methods

lift :: Quote m => DFamilyResultSig -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DFamilyResultSig -> Code m DFamilyResultSig

Lift DForallTelescope

Instance details

Methods

lift :: Quote m => DForallTelescope -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DForallTelescope -> Code m DForallTelescope

Lift DForeign

Instance details

Methods

lift :: Quote m => DForeign -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DForeign -> Code m DForeign

Lift DLetDec

Instance details

Methods

lift :: Quote m => DLetDec -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DLetDec -> Code m DLetDec

Lift DMatch

Instance details

Methods

lift :: Quote m => DMatch -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DMatch -> Code m DMatch

Lift DPat

Instance details

Methods

lift :: Quote m => DPat -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DPat -> Code m DPat

Lift DPatSynDir

Instance details

Methods

lift :: Quote m => DPatSynDir -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DPatSynDir -> Code m DPatSynDir

Lift DPragma

Instance details

Methods

lift :: Quote m => DPragma -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DPragma -> Code m DPragma

Lift DRuleBndr

Instance details

Methods

lift :: Quote m => DRuleBndr -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DRuleBndr -> Code m DRuleBndr

Lift DTySynEqn

Instance details

Methods

lift :: Quote m => DTySynEqn -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DTySynEqn -> Code m DTySynEqn

Lift DType

Instance details

Methods

lift :: Quote m => DType -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DType -> Code m DType

Lift DTypeFamilyHead

Instance details

Methods

lift :: Quote m => DTypeFamilyHead -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DTypeFamilyHead -> Code m DTypeFamilyHead

Lift NewOrData

Instance details

Methods

lift :: Quote m => NewOrData -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => NewOrData -> Code m NewOrData

Lift DFunArgs

Instance details

Methods

lift :: Quote m => DFunArgs -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DFunArgs -> Code m DFunArgs

Lift DTypeArg

Instance details

Methods

lift :: Quote m => DTypeArg -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DTypeArg -> Code m DTypeArg

Lift DVisFunArg

Instance details

Methods

lift :: Quote m => DVisFunArg -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DVisFunArg -> Code m DVisFunArg

Lift ForallTelescope

Instance details

Methods

lift :: Quote m => ForallTelescope -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => ForallTelescope -> Code m ForallTelescope

Lift FunArgs

Instance details

Methods

lift :: Quote m => FunArgs -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => FunArgs -> Code m FunArgs

Lift TypeArg

Instance details

Methods

lift :: Quote m => TypeArg -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => TypeArg -> Code m TypeArg

Lift VisFunArg

Instance details

Methods

lift :: Quote m => VisFunArg -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => VisFunArg -> Code m VisFunArg

Lift flag => Lift (DTyVarBndr flag :: Type)

Instance details

Methods

lift :: Quote m => DTyVarBndr flag -> m Exp

liftTyped :: forall (m :: Type -> Type). Quote m => DTyVarBndr flag -> Code m (DTyVarBndr flag)