Language.Haskell.TH.Desugar.OMap

Description

An OMap behaves much like a Map, with all the same asymptotics, but also remembers the order that keys were inserted.

This module offers a simplified version of the Data.Map.Ordered API that assumes left-biased indices everywhere and uses a different Semigroup instance (the one in this module uses (<>) = union) and Monoid instance (the one in this module uses mappend = union).

Documentation

newtype OMap k v

An ordered map whose insertPre, insertPost, intersection, intersectionWithKey, union, and unionWithKey operations are biased towards leftmost indices when when breaking ties between keys.

Constructors

OMap (Bias L (OMap k v))

Trivial maps

empty :: forall k v. OMap k v

singleton :: k -> v -> OMap k v

Insertion

insertPre :: Ord k => k -> v -> OMap k v -> OMap k v

The value's index will be lower than the indices of the values in the OSet.

insertPost :: Ord k => OMap k v -> k -> v -> OMap k v

The value's index will be higher than the indices of the values in the OSet.

union :: forall k v. Ord k => OMap k v -> OMap k v -> OMap k v

unionWithKey :: Ord k => (k -> v -> v -> v) -> OMap k v -> OMap k v -> OMap k v

Deletion

delete :: forall k v. Ord k => k -> OMap k v -> OMap k v

filterWithKey :: Ord k => (k -> v -> Bool) -> OMap k v -> OMap k v

(\\) :: forall k v v'. Ord k => OMap k v -> OMap k v' -> OMap k v

intersection :: forall k v v'. Ord k => OMap k v -> OMap k v' -> OMap k v

intersectionWithKey :: Ord k => (k -> v -> v' -> v'') -> OMap k v -> OMap k v' -> OMap k v''

Query

null :: forall k v. OMap k v -> Bool

size :: forall k v. OMap k v -> Int

member :: forall k v. Ord k => k -> OMap k v -> Bool

notMember :: forall k v. Ord k => k -> OMap k v -> Bool

lookup :: forall k v. Ord k => k -> OMap k v -> Maybe v

Indexing

type Index = Int

A 0-based index, much like the indices used by lists' !! operation. All indices are with respect to insertion order.

lookupIndex :: forall k v. Ord k => k -> OMap k v -> Maybe Index

lookupAt :: forall k v. Index -> OMap k v -> Maybe (k, v)

List conversions

fromList :: Ord k => [(k, v)] -> OMap k v

assocs :: forall k v. OMap k v -> [(k, v)]

toAscList :: forall k v. OMap k v -> [(k, v)]

Map conversion

toMap :: forall k v. OMap k v -> Map k v