Language.Haskell.TH.Desugar.Sweeten

Description

The functions in this module convert desugared Template Haskell back into proper Template Haskell.

Documentation

expToTH :: DExp -> Exp

matchToTH :: DMatch -> Match

patToTH :: DPat -> Pat

decsToTH :: [DDec] -> [Dec]

decToTH :: DDec -> [Dec]

This returns a list of Decs because GHC 7.6.3 does not have a one-to-one mapping between DDec and Dec.

letDecToTH :: DLetDec -> Maybe Dec

Note: This can currently only return a Nothing if the DLetDec is a pragma which is not supported by the GHC version being used.

typeToTH :: DType -> Type

conToTH :: DCon -> Con

foreignToTH :: DForeign -> Foreign

pragmaToTH :: DPragma -> Maybe Pragma

ruleBndrToTH :: DRuleBndr -> RuleBndr

clauseToTH :: DClause -> Clause

tvbToTH :: DTyVarBndr -> TyVarBndr

cxtToTH :: DCxt -> Cxt

predToTH :: DPred -> Pred

derivClauseToTH :: DDerivClause -> [DerivClause]

patSynDirToTH :: DPatSynDir -> PatSynDir

typeArgToTH :: DTypeArg -> TypeArg