Language.Haskell.TH.Desugar

Description

Desugars full Template Haskell syntax into a smaller core syntax for further processing. The desugared types and constructors are prefixed with a D.

Desugared data types

data DExp

Corresponds to TH's Exp type. Note that DLamE takes names, not patterns.

Constructors

DVarE Name

DConE Name

DLitE Lit

DAppE DExp DExp

DLamE [Name] DExp

DCaseE DExp [DMatch]

DLetE [DLetDec] DExp

DSigE DExp DType

DStaticE DExp

data DLetDec

Declarations as used in a let statement.

Constructors

DFunD Name [DClause]

DValD DPat DExp

DSigD Name DType

DInfixD Fixity Name

data DPat

Corresponds to TH's Pat type.

Constructors

DLitPa Lit

DVarPa Name

DConPa Name [DPat]

DTildePa DPat

DBangPa DPat

DWildPa

data DType

Corresponds to TH's Type type, used to represent types and kinds.

Constructors

DForallT [DTyVarBndr] DCxt DType

DAppT DType DType

DSigT DType DKind

DVarT Name

DConT Name

DArrowT

DLitT TyLit

DWildCardT

DStarT

type DKind = DType

Kinds are types.

type DCxt = [DPred]

Corresponds to TH's Cxt

data DPred

Corresponds to TH's Pred

Constructors

DAppPr DPred DType

DSigPr DPred DKind

DVarPr Name

DConPr Name

DWildCardPr

data DTyVarBndr

Corresponds to TH's TyVarBndr

Constructors

DPlainTV Name

DKindedTV Name DKind

data DMatch

Corresponds to TH's Match type.

Constructors

DMatch DPat DExp

data DClause

Corresponds to TH's Clause type.

Constructors

DClause [DPat] DExp

data DDec

Corresponds to TH's Dec type.

Constructors

DLetDec DLetDec

DDataD NewOrData DCxt Name [DTyVarBndr] [DCon] [DPred]

DTySynD Name [DTyVarBndr] DType

DClassD DCxt Name [DTyVarBndr] [FunDep] [DDec]

DInstanceD (Maybe Overlap) DCxt DType [DDec]

DForeignD DForeign

DPragmaD DPragma

DOpenTypeFamilyD DTypeFamilyHead

DClosedTypeFamilyD DTypeFamilyHead [DTySynEqn]

DDataFamilyD Name [DTyVarBndr]

DDataInstD NewOrData DCxt Name [DType] [DCon] [DPred]

DTySynInstD Name DTySynEqn

DRoleAnnotD Name [Role]

DStandaloneDerivD DCxt DType

DDefaultSigD Name DType

data Overlap :: *

Varieties of allowed instance overlap.

Constructors

Overlappable
May be overlapped by more specific instances

Overlapping
May overlap a more general instance

Overlaps
Both Overlapping and Overlappable

Incoherent
Both Overlappable and Overlappable, and pick an arbitrary one if multiple choices are available.

data NewOrData

Is it a newtype or a data type?

Constructors

Newtype

Data

data DTypeFamilyHead

Corresponds to TH's TypeFamilyHead type

Constructors

DTypeFamilyHead Name [DTyVarBndr] DFamilyResultSig (Maybe InjectivityAnn)

data DFamilyResultSig

Corresponds to TH's FamilyResultSig type

Constructors

DNoSig

DKindSig DKind

DTyVarSig DTyVarBndr

data InjectivityAnn :: *

Injectivity annotation

Constructors

InjectivityAnn Name [Name]

data DCon

Corresponds to TH's Con type.

Constructors

DCon [DTyVarBndr] DCxt Name DConFields (Maybe DType)
A GADT result type, if there is one

data DConFields

A list of fields either for a standard data constructor or a record data constructor.

Constructors

DNormalC [DBangType]

DRecC [DVarBangType]

type DBangType = (Bang, DType)

Corresponds to TH's BangType type.

type DVarBangType = (Name, Bang, DType)

Corresponds to TH's VarBangType type.

data Bang :: *

Constructors

Bang SourceUnpackedness SourceStrictness
C { {-# UNPACK #-} !}a

data SourceUnpackedness :: *

Constructors

NoSourceUnpackedness
C a

SourceNoUnpack
C { {-# NOUNPACK #-} } a

SourceUnpack
C { {-# UNPACK #-} } a

data SourceStrictness :: *

Constructors

NoSourceStrictness
C a

SourceLazy
C {~}a

SourceStrict
C {!}a

data DForeign

Corresponds to TH's Foreign type.

Constructors

DImportF Callconv Safety String Name DType

DExportF Callconv String Name DType

data DPragma

Corresponds to TH's Pragma type.

Constructors

DInlineP Name Inline RuleMatch Phases

DSpecialiseP Name DType (Maybe Inline) Phases

DSpecialiseInstP DType

DRuleP String [DRuleBndr] DExp DExp Phases

DAnnP AnnTarget DExp

DLineP Int String

data DRuleBndr

Corresponds to TH's RuleBndr type.

Constructors

DRuleVar Name

DTypedRuleVar Name DType

data DTySynEqn

Corresponds to TH's TySynEqn type (to store type family equations).

Constructors

DTySynEqn [DType] DType

data DInfo

Corresponds to TH's Info type.

Constructors

DTyConI DDec (Maybe [DInstanceDec])

DVarI Name DType (Maybe Name)
The Maybe Name stores the name of the enclosing definition (datatype, for a data constructor; class, for a method), if any

DTyVarI Name DKind

DPrimTyConI Name Int Bool
The Int is the arity; the Bool is whether this tycon is unlifted.

type DInstanceDec

Arguments

= DDec
Guaranteed to be an instance declaration

data Role :: *

Role annotations

Constructors

NominalR
nominal

RepresentationalR
representational

PhantomR
phantom

InferR
_

data AnnTarget :: *

Constructors

ModuleAnnotation

TypeAnnotation Name

ValueAnnotation Name

The Desugar class

class Desugar th ds | ds -> th where

This class relates a TH type with its th-desugar type and allows conversions back and forth. The functional dependency goes only one way because Type and Kind are type synonyms, but they desugar to different types.

Minimal complete definition

desugar, sweeten

Methods

desugar :: DsMonad q => th -> q ds

sweeten :: ds -> th

Main desugaring functions

dsExp :: DsMonad q => Exp -> q DExp

Desugar an expression

dsDecs :: DsMonad q => [Dec] -> q [DDec]

Desugar arbitrary Decs

dsType :: DsMonad q => Type -> q DType

Desugar a type

dsInfo :: DsMonad q => Info -> q DInfo

Desugar Info

dsPatOverExp :: DsMonad q => Pat -> DExp -> q (DPat, DExp)

Desugar a pattern, along with processing a (desugared) expression that is the entire scope of the variables bound in the pattern.

dsPatsOverExp :: DsMonad q => [Pat] -> DExp -> q ([DPat], DExp)

Desugar multiple patterns. Like dsPatOverExp.

dsPatX :: DsMonad q => Pat -> q (DPat, [(Name, DExp)])

Desugar a pattern, returning a list of (Name, DExp) pairs of extra variables that must be bound within the scope of the pattern

dsLetDecs :: DsMonad q => [Dec] -> q [DLetDec]

Desugar Decs that can appear in a let expression

dsTvb :: DsMonad q => TyVarBndr -> q DTyVarBndr

Desugar a TyVarBndr

dsCxt :: DsMonad q => Cxt -> q DCxt

Desugar a Cxt

dsCon :: DsMonad q => Con -> q [DCon]

Desugar a single Con.

dsForeign :: DsMonad q => Foreign -> q DForeign

Desugar a Foreign.

dsPragma :: DsMonad q => Pragma -> q DPragma

Desugar a Pragma.

dsRuleBndr :: DsMonad q => RuleBndr -> q DRuleBndr

Desugar a RuleBndr.

Secondary desugaring functions

type PatM q = WriterT [(Name, DExp)] q

Desugaring a pattern also returns the list of variables bound in as-patterns and the values they should be bound to. This variables must be brought into scope in the "body" of the pattern.

dsPred :: DsMonad q => Pred -> q DCxt

Desugar a Pred, flattening any internal tuples

dsPat :: DsMonad q => Pat -> PatM q DPat

Desugar a pattern.

dsDec :: DsMonad q => Dec -> q [DDec]

Desugar a single Dec, perhaps producing multiple DDecs

dsLetDec :: DsMonad q => Dec -> q [DLetDec]

Desugar a single Dec, perhaps producing multiple DLetDecs

dsMatches

Arguments

:: DsMonad q

=> Name
Name of the scrutinee, which must be a bare var

-> [Match]
Matches of the case statement

-> q [DMatch]

Desugar a list of matches for a case statement

dsBody

Arguments

:: DsMonad q

=> Body
body to desugar

-> [Dec]
"where" declarations

-> DExp
what to do if the guards don't match

-> q DExp

Desugar a Body

dsGuards

Arguments

:: DsMonad q

=> [(Guard, Exp)]
Guarded expressions

-> DExp
What to do if none of the guards match

-> q DExp

Desugar guarded expressions

dsDoStmts :: DsMonad q => [Stmt] -> q DExp

Desugar the Stmts in a do expression

dsComp :: DsMonad q => [Stmt] -> q DExp

Desugar the Stmts in a list or monad comprehension

dsClauses

Arguments

:: DsMonad q

=> Name
Name of the function

-> [Clause]
Clauses to desugar

-> q [DClause]

Desugar clauses to a function definition

dsBangType :: DsMonad q => BangType -> q DBangType

Desugar a BangType (or a StrictType, if you're old-fashioned)

dsVarBangType :: DsMonad q => VarBangType -> q DVarBangType

Desugar a VarBangType (or a VarStrictType, if you're old-fashioned)

dsTypeFamilyHead :: DsMonad q => TypeFamilyHead -> q DTypeFamilyHead

Desugar a TypeFamilyHead

dsFamilyResultSig :: DsMonad q => FamilyResultSig -> q DFamilyResultSig

Desugar a FamilyResultSig

Converting desugared AST back to TH AST

module Language.Haskell.TH.Desugar.Sweeten

Expanding type synonyms

expand :: (DsMonad q, Data a) => a -> q a

Expand all type synonyms and type families in the desugared abstract syntax tree provided, where type family simplification is on a "best effort" basis. Normally, the first parameter should have a type like DExp or DLetDec.

expandType :: DsMonad q => DType -> q DType

Expands all type synonyms in a desugared type. Also expands open type family applications, as long as the arguments have no free variables. Attempts to expand closed type family applications, but aborts the moment it spots anything strange, like a nested type family application or type variable.

Reification

reifyWithWarning :: Quasi q => Name -> q Info

Reify a declaration, warning the user about splices if the reify fails. The warning says that reification can fail if you try to reify a type in the same splice as it is declared.

The following definitions allow you to register a list of Decs to be used in reification queries.

withLocalDeclarations :: DsMonad q => [Dec] -> DsM q a -> q a

Add a list of declarations to be considered when reifying local declarations.

dsReify :: DsMonad q => Name -> q (Maybe DInfo)

Like reify, but safer and desugared. Uses local declarations where available.

reifyWithLocals_maybe :: DsMonad q => Name -> q (Maybe Info)

Like reify from Template Haskell, but looks also in any not-yet-typechecked declarations. To establish this list of not-yet-typechecked declarations, use withLocalDeclarations. Returns Nothing if reification fails. Note that no inferred type information is available from local declarations; bottoms may be used if necessary.

reifyWithLocals :: DsMonad q => Name -> q Info

Like reifyWithLocals_maybe, but throws an exception upon failure, warning the user about separating splices.

class Quasi m => DsMonad m where

A DsMonad stores some list of declarations that should be considered in scope. DsM is the prototypical inhabitant of DsMonad.

Minimal complete definition

localDeclarations

Methods

localDeclarations :: m [Dec]

Produce a list of local declarations.

data DsM q a

A convenient implementation of the DsMonad class. Use by calling withLocalDeclarations.

Nested pattern flattening

scExp :: DsMonad q => DExp -> q DExp

Remove all nested pattern-matches within this expression. This also removes all DTildePas and DBangPas. After this is run, every pattern is guaranteed to be either a DConPa with bare variables as arguments, a DLitPa, or a DWildPa.

scLetDec :: DsMonad q => DLetDec -> q DLetDec

Like scExp, but for a DLetDec.

Utility functions

applyDExp :: DExp -> [DExp] -> DExp

Apply one DExp to a list of arguments

applyDType :: DType -> [DType] -> DType

Apply one DType to a list of arguments

dPatToDExp :: DPat -> DExp

Convert a DPat to a DExp. Fails on DWildP.

removeWilds :: DsMonad q => DPat -> q DPat

Remove all wildcards from a pattern, replacing any wildcard with a fresh variable

getDataD

Arguments

:: Quasi q

=> String
Print this out on failure

-> Name
Name of the datatype (data or newtype) of interest

-> q ([TyVarBndr], [Con])

Extract the TyVarBndrs and constructors given the Name of a type

dataConNameToDataName :: Quasi q => Name -> q Name

From the name of a data constructor, retrive the datatype definition it is a part of.

dataConNameToCon :: Quasi q => Name -> q Con

From the name of a data constructor, retrieve its definition as a Con

nameOccursIn :: Data a => Name -> a -> Bool

Check if a name occurs anywhere within a TH tree.

allNamesIn :: Data a => a -> [Name]

Extract all Names mentioned in a TH tree.

flattenDValD :: Quasi q => DLetDec -> q [DLetDec]

If the declaration passed in is a DValD, creates new, equivalent declarations such that the DPat in all DValDs is just a plain DVarPa. Other declarations are passed through unchanged. Note that the declarations that come out of this function are rather less efficient than those that come in: they have many more pattern matches.

getRecordSelectors

Arguments

:: Quasi q

=> DType
the type of the argument

-> DCon

-> q [DLetDec]

Produces DLetDecs representing the record selector functions from the provided DCon.

mkTypeName :: Quasi q => String -> q Name

Like TH's lookupTypeName, but if this name is not bound, then we assume it is declared in the current module.

mkDataName :: Quasi q => String -> q Name

Like TH's lookupDataName, but if this name is not bound, then we assume it is declared in the current module.

newUniqueName :: Quasi q => String -> q Name

Like newName, but even more unique (unique across different splices), and with unique nameBases. Precondition: the string is a valid Haskell alphanumeric identifier (could be upper- or lower-case).

mkTupleDExp :: [DExp] -> DExp

Make a tuple DExp from a list of DExps. Avoids using a 1-tuple.

mkTupleDPat :: [DPat] -> DPat

Make a tuple DPat from a list of DPats. Avoids using a 1-tuple.

maybeDLetE :: [DLetDec] -> DExp -> DExp

If decs is non-empty, delcare them in a let:

maybeDCaseE :: String -> DExp -> [DMatch] -> DExp

If matches is non-empty, make a case statement; otherwise make an error statement

substTy :: DsMonad q => Map Name DType -> DType -> q DType

Capture-avoiding substitution on types

tupleDegree_maybe :: String -> Maybe Int

Extract the degree of a tuple

tupleNameDegree_maybe :: Name -> Maybe Int

Extract the degree of a tuple name

unboxedTupleDegree_maybe :: String -> Maybe Int

Extract the degree of an unboxed tuple

unboxedTupleNameDegree_maybe :: Name -> Maybe Int

Extract the degree of a tuple name

strictToBang :: Bang -> Bang

Convert a Strict to a Bang in GHCs 7.x. This is just the identity operation in GHC 8.x, which has no Strict. (This is included in GHC 8.x only for good Haddocking.)

Extracting bound names

extractBoundNamesStmt :: Stmt -> Set Name

Extract the names bound in a Stmt

extractBoundNamesDec :: Dec -> Set Name

Extract the names bound in a Dec that could appear in a let expression.

extractBoundNamesPat :: Pat -> Set Name

Extract the names bound in a Pat