Language.Haskell.TH.Desugar

Description

Desugars full Template Haskell syntax into a smaller core syntax for further processing.

Desugared data types

data DExp

Corresponds to TH's Exp type. Note that DLamE takes names, not patterns.

Constructors

DVarE Name

DConE Name

DLitE Lit

DAppE DExp DExp

DAppTypeE DExp DType

DLamE [Name] DExp

DCaseE DExp [DMatch]

DLetE [DLetDec] DExp

DSigE DExp DType

DStaticE DExp

data DLetDec

Declarations as used in a let statement.

Constructors

DFunD Name [DClause]

DValD DPat DExp

DSigD Name DType

DInfixD Fixity Name

DPragmaD DPragma

data DPat

Corresponds to TH's Pat type.

Constructors

DLitP Lit

DVarP Name

DConP Name [DPat]

DTildeP DPat

DBangP DPat

DSigP DPat DType

DWildP

data DType

Corresponds to TH's Type type, used to represent types and kinds.

Constructors

DForallT [DTyVarBndr] DCxt DType

DAppT DType DType

DAppKindT DType DKind

DSigT DType DKind

DVarT Name

DConT Name

DArrowT

DLitT TyLit

DWildCardT

type DKind = DType

Kinds are types. Corresponds to TH's Kind

type DCxt = [DPred]

Corresponds to TH's Cxt

type DPred = DType

Predicates are types. Corresponds to TH's Pred

data DTyVarBndr

Corresponds to TH's TyVarBndr

Constructors

DPlainTV Name

DKindedTV Name DKind

data DMatch

Corresponds to TH's Match type.

Constructors

DMatch DPat DExp

data DClause

Corresponds to TH's Clause type.

Constructors

DClause [DPat] DExp

data DDec

Corresponds to TH's Dec type.

Constructors

DLetDec DLetDec

DDataD NewOrData DCxt Name [DTyVarBndr] (Maybe DKind) [DCon] [DDerivClause]

DTySynD Name [DTyVarBndr] DType

DClassD DCxt Name [DTyVarBndr] [FunDep] [DDec]

DInstanceD (Maybe Overlap) (Maybe [DTyVarBndr]) DCxt DType [DDec]

DForeignD DForeign

DOpenTypeFamilyD DTypeFamilyHead

DClosedTypeFamilyD DTypeFamilyHead [DTySynEqn]

DDataFamilyD Name [DTyVarBndr] (Maybe DKind)

DDataInstD NewOrData DCxt (Maybe [DTyVarBndr]) DType (Maybe DKind) [DCon] [DDerivClause]

DTySynInstD DTySynEqn

DRoleAnnotD Name [Role]

DStandaloneDerivD (Maybe DDerivStrategy) (Maybe [DTyVarBndr]) DCxt DType

DDefaultSigD Name DType

DPatSynD Name PatSynArgs DPatSynDir DPat

DPatSynSigD Name DPatSynType

data DDerivClause

Corresponds to TH's DerivClause type.

Constructors

DDerivClause (Maybe DDerivStrategy) DCxt

data DDerivStrategy

Corresponds to TH's DerivStrategy type.

Constructors

DStockStrategy
A "standard" derived instance

DAnyclassStrategy
-XDeriveAnyClass

DNewtypeStrategy
-XGeneralizedNewtypeDeriving

DViaStrategy DType
-XDerivingVia

data DPatSynDir

Corresponds to TH's PatSynDir type

Constructors

DUnidir
pattern P x {<-} p

DImplBidir
pattern P x {=} p

DExplBidir [DClause]
pattern P x {<-} p where P x = e

type DPatSynType = DType

Corresponds to TH's PatSynType type

data Overlap

Varieties of allowed instance overlap.

Constructors

Overlappable
May be overlapped by more specific instances

Overlapping
May overlap a more general instance

Overlaps
Both Overlapping and Overlappable

Incoherent
Both Overlappable and Overlappable, and pick an arbitrary one if multiple choices are available.

data PatSynArgs

A pattern synonym's argument type.

Constructors

PrefixPatSyn [Name]
pattern P {x y z} = p

InfixPatSyn Name Name
pattern {x P y} = p

RecordPatSyn [Name]
pattern P { {x,y,z} } = p

data NewOrData

Is it a newtype or a data type?

Constructors

Newtype

Data

data DTypeFamilyHead

Corresponds to TH's TypeFamilyHead type

Constructors

DTypeFamilyHead Name [DTyVarBndr] DFamilyResultSig (Maybe InjectivityAnn)

data DFamilyResultSig

Corresponds to TH's FamilyResultSig type

Constructors

DNoSig

DKindSig DKind

DTyVarSig DTyVarBndr

data InjectivityAnn

Injectivity annotation

Constructors

InjectivityAnn Name [Name]

data DCon

Corresponds to TH's Con type. Unlike Con, all DCons reflect GADT syntax. This is beneficial for th-desugar's since it means that all data type declarations can support explicit return kinds, so one does not need to represent them with something like Maybe DKind, since Haskell98-style data declaration syntax isn't used. Accordingly, there are some differences between DCon and Con to keep in mind:

Unlike ForallC, where the meaning of the TyVarBndrs changes depending on whether it's followed by GadtC/RecGadtC or not, the meaning of the DTyVarBndrs in a DCon is always the same: it is the list of universally and existentially quantified type variables. Note that it is not guaranteed that one set of type variables will appear before the other.

A DCon always has an explicit return type.

Constructors

DCon [DTyVarBndr] DCxt Name DConFields DType
The GADT result type

data DConFields

A list of fields either for a standard data constructor or a record data constructor.

Constructors

DNormalC DDeclaredInfix [DBangType]

DRecC [DVarBangType]

type DDeclaredInfix = Bool

True if a constructor is declared infix. For normal ADTs, this means that is was written in infix style. For example, both of the constructors below are declared infix.

data Infix = Int Infix Int | Int :*: Int

Whereas neither of these constructors are declared infix:

data Prefix = Prefix Int Int | (:+:) Int Int

For GADTs, detecting whether a constructor is declared infix is a bit trickier, as one cannot write a GADT constructor "infix-style" like one can for normal ADT constructors. GHC considers a GADT constructor to be declared infix if it meets the following three criteria:

Its name uses operator syntax (e.g., (:*:)).

It has exactly two fields (without record syntax).

It has a programmer-specified fixity declaration.

For example, in the following GADT:

infixl 5 :**:, :&&:, :^^:, ActuallyPrefix
data InfixGADT a where
  (:**:) :: Int -> b -> InfixGADT (Maybe b) -- Only this one is infix
  ActuallyPrefix :: Char -> Bool -> InfixGADT Double
  (:&&:) :: { infixGADT1 :: b, infixGADT2 :: Int } -> InfixGADT b :: Int -> Int -> Int -> InfixGADT Int
  (:!!:) :: Char -> Char -> InfixGADT Char

Only the (:**:) constructor is declared infix. The other constructors are not declared infix, because:

ActuallyPrefix does not use operator syntax (criterion 1).

(:&&:) uses record syntax (criterion 2).

(:^^:) does not have exactly two fields (criterion 2).

(:!!:) does not have a programmer-specified fixity declaration (criterion 3).

type DBangType = (Bang, DType)

Corresponds to TH's BangType type.

type DVarBangType = (Name, Bang, DType)

Corresponds to TH's VarBangType type.

data Bang

Constructors

Bang SourceUnpackedness SourceStrictness
C { {-# UNPACK #-} !}a

data SourceUnpackedness

Constructors

NoSourceUnpackedness
C a

SourceNoUnpack
C { {-# NOUNPACK #-} } a

SourceUnpack
C { {-# UNPACK #-} } a

data SourceStrictness

Constructors

NoSourceStrictness
C a

SourceLazy
C {~}a

SourceStrict
C {!}a

data DForeign

Corresponds to TH's Foreign type.

Constructors

DImportF Callconv Safety String Name DType

DExportF Callconv String Name DType

data DPragma

Corresponds to TH's Pragma type.

Constructors

DInlineP Name Inline RuleMatch Phases

DSpecialiseP Name DType (Maybe Inline) Phases

DSpecialiseInstP DType

DRuleP String (Maybe [DTyVarBndr]) [DRuleBndr] DExp DExp Phases

DAnnP AnnTarget DExp

DLineP Int String

DCompleteP [Name] (Maybe Name)

data DRuleBndr

Corresponds to TH's RuleBndr type.

Constructors

DRuleVar Name

DTypedRuleVar Name DType

data DTySynEqn

Corresponds to TH's TySynEqn type (to store type family equations).

Constructors

DTySynEqn (Maybe [DTyVarBndr]) DType DType

data DInfo

Corresponds to TH's Info type.

Constructors

DTyConI DDec (Maybe [DInstanceDec])

DVarI Name DType (Maybe Name)
The Maybe Name stores the name of the enclosing definition (datatype, for a data constructor; class, for a method), if any

DTyVarI Name DKind

DPrimTyConI Name Int Bool
The Int is the arity; the Bool is whether this tycon is unlifted.

DPatSynI Name DPatSynType

type DInstanceDec

Arguments

= DDec
Guaranteed to be an instance declaration

data Role

Role annotations

Constructors

NominalR
nominal

RepresentationalR
representational

PhantomR
phantom

InferR
_

data AnnTarget

Constructors

ModuleAnnotation

TypeAnnotation Name

ValueAnnotation Name

The Desugar class

class Desugar th ds | ds -> th where

This class relates a TH type with its th-desugar type and allows conversions back and forth. The functional dependency goes only one way because Type and Kind are type synonyms, but they desugar to different types.

Methods

desugar :: DsMonad q => th -> q ds

sweeten :: ds -> th

Main desugaring functions

dsExp :: DsMonad q => Exp -> q DExp

Desugar an expression

dsDecs :: DsMonad q => [Dec] -> q [DDec]

Desugar arbitrary Decs

dsType :: DsMonad q => Type -> q DType

Desugar a type

dsInfo :: DsMonad q => Info -> q DInfo

Desugar Info

dsPatOverExp :: DsMonad q => Pat -> DExp -> q (DPat, DExp)

Desugar a pattern, along with processing a (desugared) expression that is the entire scope of the variables bound in the pattern.

dsPatsOverExp :: DsMonad q => [Pat] -> DExp -> q ([DPat], DExp)

Desugar multiple patterns. Like dsPatOverExp.

dsPatX :: DsMonad q => Pat -> q (DPat, [(Name, DExp)])

Desugar a pattern, returning a list of (Name, DExp) pairs of extra variables that must be bound within the scope of the pattern

dsLetDecs :: DsMonad q => [Dec] -> q ([DLetDec], DExp -> DExp)

Desugar Decs that can appear in a let expression. See the documentation for dsLetDec for an explanation of what the return type represents.

dsTvb :: DsMonad q => TyVarBndr -> q DTyVarBndr

Desugar a TyVarBndr

dsCxt :: DsMonad q => Cxt -> q DCxt

Desugar a Cxt

dsCon

Arguments

:: DsMonad q

=> [DTyVarBndr]
The universally quantified type variables (used if desugaring a non-GADT constructor).

-> DType
The original data declaration's type (used if desugaring a non-GADT constructor).

-> Con

-> q [DCon]

Desugar a single Con.

Because we always desugar Cons to GADT syntax (see the documentation for DCon), it is not always possible to desugar with just a Con alone. For instance, we must desugar:

data Foo a = forall b. MkFoo b

To this:

data Foo a :: Type where
  MkFoo :: forall a b. b -> Foo a

If our only argument was forall b. MkFoo b, it would be somewhat awkward to figure out (1) what the set of universally quantified type variables ([a]) was, and (2) what the return type (Foo a) was. For this reason, we require passing these as arguments. (If we desugar an actual GADT constructor, these arguments are ignored.)

dsForeign :: DsMonad q => Foreign -> q DForeign

Desugar a Foreign.

dsPragma :: DsMonad q => Pragma -> q DPragma

Desugar a Pragma.

dsRuleBndr :: DsMonad q => RuleBndr -> q DRuleBndr

Desugar a RuleBndr.

Secondary desugaring functions

type PatM q = WriterT [(Name, DExp)] q

Desugaring a pattern also returns the list of variables bound in as-patterns and the values they should be bound to. This variables must be brought into scope in the "body" of the pattern.

dsPred :: DsMonad q => Pred -> q DCxt

Desugar a Pred, flattening any internal tuples

dsPat :: DsMonad q => Pat -> PatM q DPat

Desugar a pattern.

dsDec :: DsMonad q => Dec -> q [DDec]

Desugar a single Dec, perhaps producing multiple DDecs

dsDataDec :: DsMonad q => NewOrData -> Cxt -> Name -> [TyVarBndr] -> Maybe Kind -> [Con] -> [DerivingClause] -> q [DDec]

Desugar a DataD or NewtypeD.

dsDataInstDec :: DsMonad q => NewOrData -> Cxt -> Name -> Maybe [TyVarBndr] -> [TypeArg] -> Maybe Kind -> [Con] -> [DerivingClause] -> q [DDec]

Desugar a DataInstD or a NewtypeInstD.

type DerivingClause = DerivClause

A backwards-compatible type synonym for the thing representing a single derived class in a deriving clause. (This is a DerivClause, Pred, or Name depending on the GHC version.)

dsDerivClause :: DsMonad q => DerivingClause -> q DDerivClause

Desugar a DerivingClause.

dsLetDec :: DsMonad q => Dec -> q ([DLetDec], DExp -> DExp)

Desugar a single Dec that can appear in a let expression. This produces the following output:

One or more DLetDecs (a single Dec can produce multiple DLetDecs in the event of a value declaration that binds multiple things by way of pattern matching.

A function of type DExp -> DExp, which should be applied to the expression immediately following the DLetDecs. This function prepends binding forms for any implicit params that were bound in the argument Dec. (If no implicit params are bound, this is simply the id function.)

For instance, if the argument to dsLetDec is the ?x = 42 part of this expression:

let { ?x = 42 } in ?x

Then the output is:

let new_x_val = 42

\z -> bindIP @"x" new_x_val z

This way, the expression let { new_x_val = 42 } in bindIP @"x" new_x_val (ip @"x") can be formed. The implicit param binders always come after all the other DLetDecs to support parallel assignment of implicit params.

dsMatches

Arguments

:: DsMonad q

=> Name
Name of the scrutinee, which must be a bare var

-> [Match]
Matches of the case statement

-> q [DMatch]

Desugar a list of matches for a case statement

dsBody

Arguments

:: DsMonad q

=> Body
body to desugar

-> [Dec]
"where" declarations

-> DExp
what to do if the guards don't match

-> q DExp

Desugar a Body

dsGuards

Arguments

:: DsMonad q

=> [(Guard, Exp)]
Guarded expressions

-> DExp
What to do if none of the guards match

-> q DExp

Desugar guarded expressions

dsDoStmts :: DsMonad q => [Stmt] -> q DExp

Desugar the Stmts in a do expression

dsComp :: DsMonad q => [Stmt] -> q DExp

Desugar the Stmts in a list or monad comprehension

dsClauses

Arguments

:: DsMonad q

=> Name
Name of the function

-> [Clause]
Clauses to desugar

-> q [DClause]

Desugar clauses to a function definition

dsBangType :: DsMonad q => BangType -> q DBangType

Desugar a BangType (or a StrictType, if you're old-fashioned)

dsVarBangType :: DsMonad q => VarBangType -> q DVarBangType

Desugar a VarBangType (or a VarStrictType, if you're old-fashioned)

dsTypeFamilyHead :: DsMonad q => TypeFamilyHead -> q DTypeFamilyHead

Desugar a TypeFamilyHead

dsFamilyResultSig :: DsMonad q => FamilyResultSig -> q DFamilyResultSig

Desugar a FamilyResultSig

dsPatSynDir :: DsMonad q => Name -> PatSynDir -> q DPatSynDir

Desugar a PatSynDir. (Available only with GHC 8.2+)

dsTypeArg :: DsMonad q => TypeArg -> q DTypeArg

Desugar a TypeArg.

Converting desugared AST back to TH AST

module Language.Haskell.TH.Desugar.Sweeten

Expanding type synonyms

expand :: (DsMonad q, Data a) => a -> q a

Expand all type synonyms and type families in the desugared abstract syntax tree provided, where type family simplification is on a "best effort" basis. Normally, the first parameter should have a type like DExp or DLetDec.

expandType :: DsMonad q => DType -> q DType

Expands all type synonyms in a desugared type. Also expands open type family applications. (In GHCs before 7.10, this part does not work if there are any variables.) Attempts to expand closed type family applications, but aborts the moment it spots anything strange, like a nested type family application or type variable.

Reification

reifyWithWarning :: (Quasi q, MonadFail q) => Name -> q Info

Reify a declaration, warning the user about splices if the reify fails. The warning says that reification can fail if you try to reify a type in the same splice as it is declared.

The following definitions allow you to register a list of Decs to be used in reification queries.

withLocalDeclarations :: DsMonad q => [Dec] -> DsM q a -> q a

Add a list of declarations to be considered when reifying local declarations.

dsReify :: DsMonad q => Name -> q (Maybe DInfo)

Like reify, but safer and desugared. Uses local declarations where available.

reifyWithLocals_maybe :: DsMonad q => Name -> q (Maybe Info)

Like reify from Template Haskell, but looks also in any not-yet-typechecked declarations. To establish this list of not-yet-typechecked declarations, use withLocalDeclarations. Returns Nothing if reification fails. Note that no inferred type information is available from local declarations; bottoms may be used if necessary.

reifyWithLocals :: DsMonad q => Name -> q Info

Like reifyWithLocals_maybe, but throws an exception upon failure, warning the user about separating splices.

reifyFixityWithLocals :: DsMonad q => Name -> q (Maybe Fixity)

Like reifyWithLocals_maybe, but for fixities. Note that a return of Nothing might mean that the name is not in scope, or it might mean that the name has no assigned fixity. (Use reifyWithLocals_maybe if you really need to tell the difference.)

lookupValueNameWithLocals :: DsMonad q => String -> q (Maybe Name)

Like lookupValueName from Template Haskell, but looks also in Names of not-yet-typechecked declarations. To establish this list of not-yet-typechecked declarations, use withLocalDeclarations. Returns Nothing if no value with the same name can be found.

lookupTypeNameWithLocals :: DsMonad q => String -> q (Maybe Name)

Like lookupTypeName from Template Haskell, but looks also in Names of not-yet-typechecked declarations. To establish this list of not-yet-typechecked declarations, use withLocalDeclarations. Returns Nothing if no type with the same name can be found.

mkDataNameWithLocals :: DsMonad q => String -> q Name

Like TH's lookupValueName, but if this name is not bound, then we assume it is declared in the current module.

Unlike mkDataName, this also consults the local declarations in scope when determining if the name is currently bound.

mkTypeNameWithLocals :: DsMonad q => String -> q Name

Like TH's lookupTypeName, but if this name is not bound, then we assume it is declared in the current module.

Unlike mkTypeName, this also consults the local declarations in scope when determining if the name is currently bound.

reifyNameSpace :: DsMonad q => Name -> q (Maybe NameSpace)

Determines a Name's NameSpace. If the NameSpace is attached to the Name itself (i.e., it is unambiguous), then that NameSpace is immediately returned. Otherwise, reification is used to lookup up the NameSpace (consulting local declarations if necessary).

Note that if a Name lives in two different NameSpaces (which can genuinely happen--for instance, mkName "==", where == is both a function and a type family), then this function will simply return whichever NameSpace is discovered first via reification. If you wish to find a Name in a particular NameSpace, use the lookupValueNameWithLocals or lookupTypeNameWithLocals functions.

class (Quasi m, MonadFail m) => DsMonad m where

A DsMonad stores some list of declarations that should be considered in scope. DsM is the prototypical inhabitant of DsMonad.

Methods

localDeclarations :: m [Dec]

Produce a list of local declarations.

data DsM q a

A convenient implementation of the DsMonad class. Use by calling withLocalDeclarations.

Nested pattern flattening

scExp :: DsMonad q => DExp -> q DExp

Remove all nested pattern-matches within this expression. This also removes all DTildePas and DBangPas. After this is run, every pattern is guaranteed to be either a DConPa with bare variables as arguments, a DLitPa, or a DWildPa.

scLetDec :: DsMonad q => DLetDec -> q DLetDec

Like scExp, but for a DLetDec.

Capture-avoiding substitution and utilities

module Language.Haskell.TH.Desugar.Subst

Free variable calculation

fvDType :: DType -> OSet Name

Compute the free variables of a DType.

extractBoundNamesDPat :: DPat -> OSet Name

Extract the term variables bound by a DPat.

This does not extract any type variables bound by pattern signatures.

Utility functions

applyDExp :: DExp -> [DExp] -> DExp

Apply one DExp to a list of arguments

dPatToDExp :: DPat -> DExp

Convert a DPat to a DExp. Fails on DWildP.

removeWilds :: DsMonad q => DPat -> q DPat

Remove all wildcards from a pattern, replacing any wildcard with a fresh variable

getDataD

Arguments

:: DsMonad q

=> String
Print this out on failure

-> Name
Name of the datatype (data or newtype) of interest

-> q ([TyVarBndr], [Con])

Extract the TyVarBndrs and constructors given the Name of a type

dataConNameToDataName :: DsMonad q => Name -> q Name

From the name of a data constructor, retrive the datatype definition it is a part of.

dataConNameToCon :: DsMonad q => Name -> q Con

From the name of a data constructor, retrieve its definition as a Con

nameOccursIn :: Data a => Name -> a -> Bool

Check if a name occurs anywhere within a TH tree.

allNamesIn :: Data a => a -> [Name]

Extract all Names mentioned in a TH tree.

flattenDValD :: Quasi q => DLetDec -> q [DLetDec]

If the declaration passed in is a DValD, creates new, equivalent declarations such that the DPat in all DValDs is just a plain DVarPa. Other declarations are passed through unchanged. Note that the declarations that come out of this function are rather less efficient than those that come in: they have many more pattern matches.

getRecordSelectors

Arguments

:: DsMonad q

=> DType
the type of the argument

-> [DCon]

-> q [DLetDec]

Produces DLetDecs representing the record selector functions from the provided DCons.

Note that if the same record selector appears in multiple constructors, getRecordSelectors will return only one binding for that selector. For example, if you had:

data X = X1 {y :: Symbol} | X2 {y :: Symbol}

Then calling getRecordSelectors on [X1, X2] will return:

[ DSigD y (DAppT (DAppT DArrowT (DConT X)) (DConT Symbol))
, DFunD y [ DClause [DConP X1 [DVarP field]] (DVarE field)
          , DClause [DConP X2 [DVarP field]] (DVarE field) ] ]

instead of returning one binding for X1 and another binding for X2.

getRecordSelectors attempts to filter out "naughty" record selectors whose types mention existentially quantified type variables. But see the documentation for conExistentialTvbs for limitations to this approach.

mkTypeName :: Quasi q => String -> q Name

Like TH's lookupTypeName, but if this name is not bound, then we assume it is declared in the current module.

mkDataName :: Quasi q => String -> q Name

Like TH's lookupDataName, but if this name is not bound, then we assume it is declared in the current module.

newUniqueName :: Quasi q => String -> q Name

Like newName, but even more unique (unique across different splices), and with unique nameBases. Precondition: the string is a valid Haskell alphanumeric identifier (could be upper- or lower-case).

mkTupleDExp :: [DExp] -> DExp

Make a tuple DExp from a list of DExps. Avoids using a 1-tuple.

mkTupleDPat :: [DPat] -> DPat

Make a tuple DPat from a list of DPats. Avoids using a 1-tuple.

maybeDLetE :: [DLetDec] -> DExp -> DExp

If decs is non-empty, delcare them in a let:

maybeDCaseE :: String -> DExp -> [DMatch] -> DExp

If matches is non-empty, make a case statement; otherwise make an error statement

mkDLamEFromDPats :: DsMonad q => [DPat] -> DExp -> q DExp

Convert a list of DPat arguments and a DExp body into a DLamE. This is needed since DLamE takes a list of Names for its bound variables instead of DPats, so some reorganization is needed.

tupleDegree_maybe :: String -> Maybe Int

Extract the degree of a tuple

tupleNameDegree_maybe :: Name -> Maybe Int

Extract the degree of a tuple name

unboxedSumDegree_maybe :: String -> Maybe Int

Extract the degree of an unboxed sum

unboxedSumNameDegree_maybe :: Name -> Maybe Int

Extract the degree of an unboxed sum name

unboxedTupleDegree_maybe :: String -> Maybe Int

Extract the degree of an unboxed tuple

unboxedTupleNameDegree_maybe :: Name -> Maybe Int

Extract the degree of an unboxed tuple name

strictToBang :: Bang -> Bang

Convert a Strict to a Bang in GHCs 7.x. This is just the identity operation in GHC 8.x, which has no Strict. (This is included in GHC 8.x only for good Haddocking.)

isTypeKindName :: Name -> Bool

Returns True if the argument Name is that of Type (or * or ★, to support older GHCs).

typeKindName :: Name

The Name of:

The kind Type, on GHC 8.0 or later.

The kind * on older GHCs.

bindIP :: forall name a r. a -> (IP name a => r) -> r

Get an implicit param constraint (IP name a, which is the desugared form of (?name :: a)) from an explicit value.

This function is only available with GHC 8.0 or later.

unravel :: DType -> ([DTyVarBndr], [DPred], [DType], DType)

Decompose a function type into its type variables, its context, its argument types, and its result type.

conExistentialTvbs

Arguments

:: DsMonad q

=> DType
The type of the original data declaration

-> DCon

-> q [DTyVarBndr]

Returns all of a constructor's existentially quantified type variable binders.

Detecting the presence of existentially quantified type variables in the context of Template Haskell is quite involved. Here is an example that we will use to explain how this works:

data family Foo a b
data instance Foo (Maybe a) b where
  MkFoo :: forall x y z. x -> y -> z -> Foo (Maybe x) [z]

In MkFoo, x is universally quantified, whereas y and z are existentially quantified. Note that MkFoo desugars (in Core) to something like this:

data instance Foo (Maybe a) b where
  MkFoo :: forall a b y z. (b ~ [z]). a -> y -> z -> Foo (Maybe a) b

Here, we can see that a appears in the desugared return type (it is a simple alpha-renaming of x), so it is universally quantified. On the other hand, neither y nor z appear in the desugared return type, so they are existentially quantified.

This analysis would not have been possible without knowing what the original data declaration's type was (in this case, Foo (Maybe a) b), which is why we require it as an argument. Our algorithm for detecting existentially quantified variables is not too different from what was described above: we match the constructor's return type with the original data type, forming a substitution, and check which quantified variables are not part of the domain of the substitution.

Be warned: this may overestimate which variables are existentially quantified when kind variables are involved. For instance, consider this example:

data S k (a :: k)
data T a where
  MkT :: forall k (a :: k). { foo :: Proxy (a :: k), bar :: S k a } -> T a

Here, the kind variable k does not appear syntactically in the return type T a, so conExistentialTvbs would mistakenly flag k as existential.

There are various tricks we could employ to improve this, but ultimately, making this behave correctly with respect to PolyKinds 100% of the time would amount to performing kind inference in Template Haskell, which is quite difficult. For the sake of simplicity, we have decided to stick with a dumb-but-predictable syntactic check.

mkExtraDKindBinders :: DsMonad q => DKind -> q [DTyVarBndr]

Create new kind variable binder names corresponding to the return kind of a data type. This is useful when you have a data type like:

data Foo :: forall k. k -> Type -> Type where ...

But you want to be able to refer to the type Foo a b. mkExtraDKindBinders will take the kind forall k. k -> Type -> Type, discover that is has two visible argument kinds, and return as a result two new kind variable binders [a :: k, b :: Type], where a and b are fresh type variable names.

This expands kind synonyms if necessary.

dTyVarBndrToDType :: DTyVarBndr -> DType

Convert a DTyVarBndr into a DType

toposortTyVarsOf :: [DType] -> [DTyVarBndr]

Take a list of DTypes, find their free variables, and sort them in reverse topological order to ensure that they are well scoped. In other words, the free variables are ordered such that:

Whenever an explicit kind signature of the form (A :: K) is encountered, the free variables of K will always appear to the left of the free variables of A in the returned result.

The constraint in (1) notwithstanding, free variables will appear in left-to-right order of their original appearance.

On older GHCs, this takes measures to avoid returning explicitly bound kind variables, which was not possible before TypeInType.

TypeArg

data TypeArg

An argument to a type, either a normal type (TANormal) or a visible kind application (TyArg).

TypeArg is useful when decomposing an application of a Type to its arguments (e.g., in unfoldType).

Constructors

TANormal Type

TyArg Kind

applyType :: Type -> [TypeArg] -> Type

Apply one Type to a list of arguments.

filterTANormals :: [TypeArg] -> [Type]

Filter the normal type arguments from a list of TypeArgs.

unfoldType :: Type -> (Type, [TypeArg])

Decompose an applied type into its individual components. For example, this:

Proxy @Type Char

would be unfolded to this:

(ConT ''Proxy, [TyArg (ConT ''Type), TANormal (ConT ''Char)])

DTypeArg

data DTypeArg

An argument to a type, either a normal type (DTANormal) or a visible kind application (DTyArg).

DTypeArg does not appear directly in the th-desugar AST, but it is useful when decomposing an application of a DType to its arguments.

Constructors

DTANormal DType

DTyArg DKind

applyDType :: DType -> [DTypeArg] -> DType

Apply one DType to a list of arguments

filterDTANormals :: [DTypeArg] -> [DType]

Filter the normal type arguments from a list of DTypeArgs.

unfoldDType :: DType -> (DType, [DTypeArg])

Decompose an applied type into its individual components. For example, this:

Proxy @Type Char

would be unfolded to this:

(DConT ''Proxy, [DTyArg (DConT ''Type), DTANormal (DConT ''Char)])

Extracting bound names

extractBoundNamesStmt :: Stmt -> OSet Name

Extract the names bound in a Stmt

extractBoundNamesDec :: Dec -> OSet Name

Extract the names bound in a Dec that could appear in a let expression.

extractBoundNamesPat :: Pat -> OSet Name

Extract the names bound in a Pat