Language.Haskell.TH.Desugar.Lift

Description

Defines Lift instances for the desugared language. This is defined in a separate module because it also must define Lift instances for several TH types, which are orphans and may want another definition downstream.

Orphan instances

Lift DTySynEqn

Methods

lift :: DTySynEqn -> Q Exp

Lift DRuleBndr

Methods

lift :: DRuleBndr -> Q Exp

Lift DPragma

Methods

lift :: DPragma -> Q Exp

Lift DForeign

Methods

lift :: DForeign -> Q Exp

Lift DConFields

Methods

lift :: DConFields -> Q Exp

Lift DCon

Methods

lift :: DCon -> Q Exp

Lift DFamilyResultSig

Methods

lift :: DFamilyResultSig -> Q Exp

Lift DTypeFamilyHead

Methods

lift :: DTypeFamilyHead -> Q Exp

Lift DDec

Methods

lift :: DDec -> Q Exp

Lift NewOrData

Methods

lift :: NewOrData -> Q Exp

Lift DLetDec

Methods

lift :: DLetDec -> Q Exp

Lift DClause

Methods

lift :: DClause -> Q Exp

Lift DMatch

Methods

lift :: DMatch -> Q Exp

Lift DTyVarBndr

Methods

lift :: DTyVarBndr -> Q Exp

Lift DPred

Methods

lift :: DPred -> Q Exp

Lift DType

Methods

lift :: DType -> Q Exp

Lift DPat

Methods

lift :: DPat -> Q Exp

Lift DExp

Methods

lift :: DExp -> Q Exp