Language.Haskell.TH.Desugar.Lift

Description

Defines Lift instances for the desugared language. This is defined in a separate module because it also must define Lift instances for several TH types, which are orphans and may want another definition downstream.

Orphan instances

Lift TypeArg

Instance details

Methods

lift :: TypeArg -> Q Exp

liftTyped :: TypeArg -> Q (TExp TypeArg)

Lift VisFunArg

Instance details

Methods

lift :: VisFunArg -> Q Exp

liftTyped :: VisFunArg -> Q (TExp VisFunArg)

Lift FunArgs

Instance details

Methods

lift :: FunArgs -> Q Exp

liftTyped :: FunArgs -> Q (TExp FunArgs)

Lift ForallVisFlag

Instance details

Methods

lift :: ForallVisFlag -> Q Exp

liftTyped :: ForallVisFlag -> Q (TExp ForallVisFlag)

Lift DDerivStrategy

Instance details

Methods

lift :: DDerivStrategy -> Q Exp

liftTyped :: DDerivStrategy -> Q (TExp DDerivStrategy)

Lift DDerivClause

Instance details

Methods

lift :: DDerivClause -> Q Exp

liftTyped :: DDerivClause -> Q (TExp DDerivClause)

Lift DTySynEqn

Instance details

Methods

lift :: DTySynEqn -> Q Exp

liftTyped :: DTySynEqn -> Q (TExp DTySynEqn)

Lift DRuleBndr

Instance details

Methods

lift :: DRuleBndr -> Q Exp

liftTyped :: DRuleBndr -> Q (TExp DRuleBndr)

Lift DPragma

Instance details

Methods

lift :: DPragma -> Q Exp

liftTyped :: DPragma -> Q (TExp DPragma)

Lift DForeign

Instance details

Methods

lift :: DForeign -> Q Exp

liftTyped :: DForeign -> Q (TExp DForeign)

Lift DConFields

Instance details

Methods

lift :: DConFields -> Q Exp

liftTyped :: DConFields -> Q (TExp DConFields)

Lift DCon

Instance details

Methods

lift :: DCon -> Q Exp

liftTyped :: DCon -> Q (TExp DCon)

Lift DFamilyResultSig

Instance details

Methods

lift :: DFamilyResultSig -> Q Exp

liftTyped :: DFamilyResultSig -> Q (TExp DFamilyResultSig)

Lift DTypeFamilyHead

Instance details

Methods

lift :: DTypeFamilyHead -> Q Exp

liftTyped :: DTypeFamilyHead -> Q (TExp DTypeFamilyHead)

Lift DPatSynDir

Instance details

Methods

lift :: DPatSynDir -> Q Exp

liftTyped :: DPatSynDir -> Q (TExp DPatSynDir)

Lift DDec

Instance details

Methods

lift :: DDec -> Q Exp

liftTyped :: DDec -> Q (TExp DDec)

Lift NewOrData

Instance details

Methods

lift :: NewOrData -> Q Exp

liftTyped :: NewOrData -> Q (TExp NewOrData)

Lift DLetDec

Instance details

Methods

lift :: DLetDec -> Q Exp

liftTyped :: DLetDec -> Q (TExp DLetDec)

Lift DClause

Instance details

Methods

lift :: DClause -> Q Exp

liftTyped :: DClause -> Q (TExp DClause)

Lift DMatch

Instance details

Methods

lift :: DMatch -> Q Exp

liftTyped :: DMatch -> Q (TExp DMatch)

Lift DTyVarBndr

Instance details

Methods

lift :: DTyVarBndr -> Q Exp

liftTyped :: DTyVarBndr -> Q (TExp DTyVarBndr)

Lift DType

Instance details

Methods

lift :: DType -> Q Exp

liftTyped :: DType -> Q (TExp DType)

Lift DPat

Instance details

Methods

lift :: DPat -> Q Exp

liftTyped :: DPat -> Q (TExp DPat)

Lift DExp

Instance details

Methods

lift :: DExp -> Q Exp

liftTyped :: DExp -> Q (TExp DExp)

Lift DVisFunArg

Instance details

Methods

lift :: DVisFunArg -> Q Exp

liftTyped :: DVisFunArg -> Q (TExp DVisFunArg)

Lift DFunArgs

Instance details

Methods

lift :: DFunArgs -> Q Exp

liftTyped :: DFunArgs -> Q (TExp DFunArgs)

Lift DTypeArg

Instance details

Methods

lift :: DTypeArg -> Q Exp

liftTyped :: DTypeArg -> Q (TExp DTypeArg)