Language.Haskell.TH.Desugar.Lift

Description

Defines Lift instances for the desugared language. This is defined in a separate module because it also must define Lift instances for several TH types, which are orphans and may want another definition downstream.

Orphan instances

Lift DDerivStrategy

Instance details

Methods

lift :: DDerivStrategy -> Q Exp

Lift DDerivClause

Instance details

Methods

lift :: DDerivClause -> Q Exp

Lift DTySynEqn

Instance details

Methods

lift :: DTySynEqn -> Q Exp

Lift DRuleBndr

Instance details

Methods

lift :: DRuleBndr -> Q Exp

Lift DPragma

Instance details

Methods

lift :: DPragma -> Q Exp

Lift DForeign

Instance details

Methods

lift :: DForeign -> Q Exp

Lift DConFields

Instance details

Methods

lift :: DConFields -> Q Exp

Lift DCon

Instance details

Methods

lift :: DCon -> Q Exp

Lift DFamilyResultSig

Instance details

Methods

lift :: DFamilyResultSig -> Q Exp

Lift DTypeFamilyHead

Instance details

Methods

lift :: DTypeFamilyHead -> Q Exp

Lift DPatSynDir

Instance details

Methods

lift :: DPatSynDir -> Q Exp

Lift DDec

Instance details

Methods

lift :: DDec -> Q Exp

Lift NewOrData

Instance details

Methods

lift :: NewOrData -> Q Exp

Lift DLetDec

Instance details

Methods

lift :: DLetDec -> Q Exp

Lift DClause

Instance details

Methods

lift :: DClause -> Q Exp

Lift DMatch

Instance details

Methods

lift :: DMatch -> Q Exp

Lift DTyVarBndr

Instance details

Methods

lift :: DTyVarBndr -> Q Exp

Lift DType

Instance details

Methods

lift :: DType -> Q Exp

Lift DPat

Instance details

Methods

lift :: DPat -> Q Exp

Lift DExp

Instance details

Methods

lift :: DExp -> Q Exp

Lift TypeArg

Instance details

Methods

lift :: TypeArg -> Q Exp

Lift DTypeArg

Instance details

Methods

lift :: DTypeArg -> Q Exp